requirements.txt
google_api_python_client
google_auth_oauthlib
gunicorn
langchain-core==0.2.22
packaging==23.2
optuna
pandas
protobuf
python-dotenv
pyspark
notebook==7.3.0
jupyterlab==4.3.8
python_dateutil
rapidfuzz
scikit_learn
sentence_transformers
torch
tqdm
numpy
jupyter
llama-cpp-python
Flask